# Link-free redesign