--- a/src/green.jl
+++ b/src/green.jl
@@ -1,3 +1,0 @@
-struct GreenFunction{T,N,L,NL}
-    bandstructure::Bandstructure{T,N,L,NL}
-end