--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,4 +1,4 @@
 include("test_lattice.jl")
-include("test_model.jl")
+# include("test_model.jl")
 include("test_system.jl")
-include("test_bandstructure.jl")
+# include("test_bandstructure.jl")