This is a comprehensive redesign of the `Elsa` internals and API. It addresses some of the comments in #3. 

## The problem(s)
The design in current master is based in a clear separation between linking of sites (as a geometric property) and the hoppings that are applied to said links (as a physical property). This has a number of drawbacks that became clear by considering the comments in #3 by @Datseris. 
- One of them is the difficulty to specify different hoppings at different links that differ on, say, a given intercell distance. The only way was to make a function of two positions that would include a condition on said distance, but it's not the most elegant approach. 
- Another problem is the redundancy between the stored links (of type `Links`, innternal to `Lattice`) and the Hamiltonian in `System`. These contain quite redundant information, although for some problems (say, deforming a lattice), the information in `Links` was easier to manipulate. 
- Apart from the redundancy, and the associated memory footprint, there is the performance of building a large system. Both the `Links` and the Hamiltonian had to be constructed, which leaves at least a factor of ~2 in performance on the table.
- One more issue raised, which is debatable but I think I agree with, is that the use of types for all possible directives is not the most idiomatic/Julian approach. It had the advantage of allowing building a system in a way that depended on the applied directives. However, there is a much more Julian way, which is to chain a series of operations on a given system/lattice. This is the approach that this PR now explores

## The proposed solution
This PR excises all the `Links` machinery, and stores all the information of the lattice connectivity directly as hoppings in a Hamiltonian. As such, linking two sites is just a matter of adding a hopping between them to the Hamiltonian. This is similar to the Kwant and Pybinding approach, I believe.

As a result there is no longer a distinction between `Lattice` and `System`. So `Lattice` is gone. A `System` is just a collection of sublattices, a Bravais matrix and a Hamiltonian (of type `Operator`). 
Systems are build with simple constructors that take either
1- `Sublat`s, `Bravais` vectors and a `Model`
2- A preset, and optionally a `Model` to be applied to the preset
3- Another `System` and a `Model` to replace the Hamiltonian of the original system.

In addition to this constructor, the System API includes a number of functions, that can be applied sequentially in any given order to build derived systems from one or more systems. They allow piping with `|>` for cleaner syntax. These functions are `modifysublats!`, `grow`, `transform`, `transform!` and `combine` (for the moment). We also have the function `hamiltonian` to extract the Hamiltonian of a system. They are all documented, both with docstrings and in the new [Getting Started doc](https://github.com/pablosanjose/Elsa.jl/blob/design_v3/docs/src/index.md).

One problem I had with this PR is that it required almost a full rewrite of the Elsa basics, and as a result the work is not yet at feature-parity with master. This PR allows to build Hamiltonians, and to visualise systems with the new PR in ElsaPlots.jl, but bandstructure calculations (to be renamed "Spectrum") is not yet done.

Performance-wise this PR is a step forward. Building a Hamiltonian from scratch is now approximately 1.5x to 2x as fast as in master, and uses less than half as much memory. Regarding the new syntax, I think I like it better than before, although it is less compact, but I would appreciate some feedback if anybody can provide some.

Some runtime/syntax comparisons (circular graphene flake of radius `100a0`, nearest-neighbor hopping)

Current master:
```
julia> @time using Elsa
  0.853706 seconds (1.80 M allocations: 114.473 MiB, 8.07% gc time)

julia> test() = System(Lattice(:honeycomb, LinkRule(1/√3), Region(:circle, 100)), Model(Hopping(1)))

julia> using BenchmarkTools; @btime test()
  36.359 ms (2176 allocations: 43.33 MiB)
System{Float64,2,0} : 0D system in 2D space with Float64 sites.
    Bravais vectors : ()
    Number of sites : 72562
    Sublattice names : (:A, :B)
    Unique Links : 108445
    Model with sublattice site dimensions () (default 1)
    Bloch operator of dimensions (72562, 72562) with 216890 elements
```
This PR
```
julia> @time using Elsa
  0.803117 seconds (1.74 M allocations: 111.029 MiB, 9.03% gc time)

julia> test() = System("honeycomb", Model(hopping(1, range = 1/√3))) |> grow(region = Region("circle", 100));

julia> using BenchmarkTools; @btime test()
  20.890 ms (946 allocations: 13.41 MiB)
System{Complex{Float64},Float64,2,0} : 0D system in 2D space
  Bravais vectors     : ()
  Sublattice names    : ("A", "B")
  Sublattice orbitals : (1, 1)
  Total sites         : 72562 [Float64]
  Total hoppings      : 216890 [Complex{Float64}]
  Coordination        : 2.989030070835975
```

CC @Datseris