@Datseris 2019-02-22
Hi Pablo!

> This PR excises all the `Links` machinery, and stores all the information of the lattice connectivity directly as hoppings in a Hamiltonian. As such, linking two sites is just a matter of adding a hopping between them to the Hamiltonian. This is similar to the Kwant and Pybinding approach, I believe.

Yes that is correct.

---

I read through this PR message. I very happy that you got almost ~2 performance benefit (which for tight binding applicaitons is _huge_ as they "human measurable" time, i.e. more tha miliseconds). Also happy to see that the core types necessary to get going are just 3: namely the sublattices, the Bravais vectors and the Model. This sounds very intuitive to me, and if we strip the process to the bare basics this is the minimum amount of information one needs! it is also nice to see taking advantage of the pipe syntax. Most users coming from Python etc. are unfamiliar with it of course, but I believe it is worth it to take "risks" let's say that may lead to a nicer interface. I personally like your approach _even though I never use the pipe syntax personally_.

I haven't read the documentation yet. I will do that and will give you more feedback!