@Datseris 2019-02-22
Ah wait I am a bit confused. Why is it that the example has:
```
Bravais vectors     : ()
```
? Shouldn't it list the actual vectors?