@pablosanjose 2019-02-22
You mean example 2, yes? That is a graphene flake of a given radius, so it is not an infinite Bravais lattice. It has no translational invariance, and no Bravais vectors. It is very different from `System("honeycomb")`, which is truly an infinite system with two Bravais vectors.