@Datseris 2019-02-23
Alright I am going through the docs now! You should consider to answer all the questions I ask here in the documentation document, to not do double effort!

---

> hopping functions take the position r of the center of the link and its vector dr (h = (r,dr) -> hfunc(r,dr))

What does "center of link" mean as `r`? Is it the position of Site A while `dr` is the difference between A and B? If yes then the way it is phrased in the docs is unclear. The way I read it in the docs is that r (as a vector) is r = (r_A + r_B)/2 which is certainly unintuitive.

---

Is it mandatory to use a specific type for the sublattice naming scheme, e.g. only Strings or only Symbols?

Or one can use anything?

---

> ndists n1, n2... in hopping are integer vectors or tuples that define the Bravais distance between two unit cells to be linked. In other words, to link any given unit cell with another at distance bravais * n1 specify n1 in ndists.
> range is the maximum Euclidean distance (1 by default) at which to look for neighbor candidates of each site.

This is a bit confusing for me . So `n1, n2` specify only _exact_ _specific_ pairs while `range` is saying "just do all the pairs within this distance". ? A bit more discussion about these two arguments and their difference is in order.

---

The default values of `Model` should be exposed more, and more clearly I believe. (See below where I ask the value of onsite entries).

---

Does `transform!` keep the hoppings as is?

---

On the first example:

> Create another sublattice :B, again with a single site, but at (0.0, 0.5/sqrt(3.0)

Typo here, because you use `"B"` instead of `:B` in the code snippet.

> model = Model(hopping(1, range = a0/√3));

What is the value of the onsite elements? By default zero? How does the "Model" know when to stop adding onsite elements? This instruction only has info about the hopping.  This has confused me a bit. Does `Model` always correspond to an infinite system that must be made finite afterwards? 

From your comments on this PR and from the documentation I think it is clear that now there is no such thing as "lattice". A system is instantiated when and only when the Hamiltonian of the system gets an entry at some position. 

In addition the docs say:

>Model: A tight-binding model given as a set of hopping and onsite instructions Model(onsite(...), hopping(...), ...), with the following syntax

So I guess this expression:

> System("honeycomb") |> grow(region = Region("circle", 3))

Just uses the already existing functions in `System` to "grow" it? Where does it get the onsite values from? 

But still, the most important question is:

**As it seems, the first instance of `System` a user can create is always infinite and you must "grow" it to get a finite system** . Correct? If yes, then adding this exact statement in the docs could be useful.


---

For later analysis, is it possible to still obtain the "original" Bravais vectors from a finite system?

---

The "presets" part can be improved I think. It is not very good practice to have different presets (all of type string) give different type of output (of course this is totally personal opinion). You mention that some presets do and some do not have a "model". After reading the presets parts, it is unclear to me what a preset is.

---

## Conclusion

This looks seriously good. I actually like the piping syntax. But it should be made more clear what is finite and what is infinite I believe. The name `grow` is truly opposing its name. `grow` takes an **infinite system** and makes it **finite**. So certainly the system doesn't "grow" in reality :P 

And in general I think this:

> grow(system; supercell, region) to expand the unit cell of a system with Bravais vectors bravais so that the result has bravais vectors bravais * supercell (could be less vectors than the original). region is a boolean function of spatial position that specifies if a site at r in the unit cell should be included in the result.

Can be a bit confusing for a newcomer. Shouldn't it be simple to state how to make a "finite" system? This involves the concept of Super cells, and the product of a bravais vector with a supercell. I admit that I myself do not immediatelly understand what it means. But I do immediatelly understand "hey this is a finite graphene system looking like a circle!".

lastly, as a personal comment:

```
julia> grow(sys, region = incircle)
System{Complex{Float64},Float64,2,0} : 0D system in 2D space
  Bravais vectors     : ()
```
I will never find intuitive the choice of not displaying the Bravais vectors. Yes, on paper you are absolutely true, there are no "true Bravais vectors", but still there _are_ the vectors that connect the unit cells with each other and these are invariant! You already state clearly that the system is `0D`, in the first line. Is there no way to persuade you to just normally display the Bravais vectors as well?